+++ src/game/Unit.cpp
@@ -2282,7 +2282,7 @@ void Unit::CalculateDamageAbsorbAndResist(Unit *pCaster, DamageInfo* damageInfo,
 
     // for absorb use only absorb_affected_damage
     uint32 absorb_affected_damage = pCaster->CalcNotIgnoreAbsorbDamage(damageInfo);
-    uint32 absorb_unaffected_damage = RemainingDamage > absorb_affected_damage ?
+    uint32 absorb_unaffected_damage = RemainingDamage > int32(absorb_affected_damage) ?
                                       RemainingDamage - absorb_affected_damage : 0;
 
     RemainingDamage -= absorb_unaffected_damage;
@@ -7961,7 +7961,7 @@ void Unit::SpellDamageBonusDone(DamageInfo* damageInfo, uint32 stack)
     }
     else
     {
-        damageInfo->bonusDone  = - damageInfo->damage;
+        damageInfo->bonusDone  = -int32(damageInfo->damage);
         damageInfo->damage     = 0;
     }
 }
@@ -8029,7 +8029,7 @@ void Unit::SpellDamageBonusTaken(DamageInfo* damageInfo, uint32 stack)
     }
     else
     {
-        damageInfo->bonusTaken = -damageInfo->damage;
+        damageInfo->bonusTaken = -int32(damageInfo->damage);
         damageInfo->damage     = 0;
     }
 }
@@ -9041,7 +9041,7 @@ void Unit::MeleeDamageBonusDone(DamageInfo* damageInfo, uint32 stack)
     }
     else
     {
-        damageInfo->bonusDone = -damageInfo->damage;
+        damageInfo->bonusDone = -int32(damageInfo->damage);
         damageInfo->damage     = 0;
     }
 }
@@ -9141,7 +9141,7 @@ void Unit::MeleeDamageBonusTaken(DamageInfo* damageInfo, uint32 stack)
     }
     else
     {
-        damageInfo->bonusTaken = -damageInfo->damage;
+        damageInfo->bonusTaken = -int32(damageInfo->damage);
         damageInfo->damage     = 0;
     }
 }
